% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishRmap.R
\name{fishRmap}
\alias{fishRmap}
\title{Call visualization function}
\usage{
fishRmap(userdata)
}
\arguments{
\item{userdata}{A string path to a directory file (eventually R vars instead/in addition, but for now we use dir to 2 text files)}
}
\value{
TBD
}
\description{
Call visualization function
}
\examples{
someExample <- 'goes here' 
}

