% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishTrade.R
\name{fishTrade}
\alias{fishTrade}
\title{Merge user data with shapefile data to get a spatial object with a very wide data socket}
\usage{
fishTrade(userDT, world)
}
\arguments{
\item{userDT}{A data.table returned from fishReader}

\item{world}{A spatial object returned from worldEater}
}
\value{
A data.table containing column with line graphs to be appended to leaflet
}
\description{
Merge user data with shapefile data to get a spatial object with a very wide data socket
}
\examples{
someExample <- 'goes here' 
}

