% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinFishWorld.R
\name{joinFishWorld}
\alias{joinFishWorld}
\title{Merge user data with shapefile data to get a spatial object with a very wide data socket}
\usage{
joinFishWorld(userDT, world)
}
\arguments{
\item{userDT}{A data.table returned from fishReader}

\item{world}{A spatial object returned from worldEater}
}
\value{
An object of class 'SpatialLinesDataFrame' to be appended to leaflet
}
\description{
Merge user data with shapefile data to get a spatial object with a very wide data socket
}
\examples{
someExample <- 'goes here' 
}

