% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parser}
\alias{parser}
\title{Parse text strings. Why bother? Because we do.call it in fishingLines}
\usage{
parser(someStr)
}
\arguments{
\item{someStr}{A character string that, when parsed, returns an object}
}
\value{
An object (for our purposes, a single SpatialLine or gcIntermediate line, but may use elsewhere)
}
\description{
Parse text strings. Why bother? Because we do.call it in fishingLines
}
\examples{
someExample <- 'goes here' 
}

