% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldEater.R
\name{worldEater}
\alias{worldEater}
\title{Read (and, if needed, first get and extract) world file. Note: If we include shapefile with library, this may be unnecessary}
\usage{
worldEater()
}
\value{
A spatial object to be appended to leaflet
}
\description{
Read (and, if needed, first get and extract) world file. Note: If we include shapefile with library, this may be unnecessary
}
\examples{
someExample <- 'goes here' 
}

